requirements.txt
python-dotenv
langchain==0.2.15
langchain-openai==0.1.23
langchain-community==0.2.15
chromadb==1.2.1
pymupdf
reportlab
